GETP-222/refactor: 홈페이지 디렉토리 구조 변경
