+++ src/Router.tsx
@@ -6,8 +6,8 @@ import SignUpPage from "@/pages/auth/SignUpPage";
 import EditClientPage from "@/pages/client/EditClientPage";
 import MeetingRequestPage from "@/pages/client/MeetingRequest/MeetingRequestPage";
 import RegisterClientPage from "@/pages/client/RegisterClientPage";
+import NotFoundPage from "@/pages/error/NotFoundPage";
 import HomePage from "@/pages/home/HomePage";
-import NotFoundPage from "@/pages/home/NotFoundPage";
 import PeopleDetailPage from "@/pages/people/PeopleDetail/PeopleDetailPage";
 import PeopleInfoRegisterPage from "@/pages/people/PeopleInfoRegister/PeopleInfoRegisterPage";
 import PeopleListPage from "@/pages/people/PeopleList/PeopleListPage";
