+++ src/components/home/Carousel/Carousel.style.ts
