+++ src/components/home/Carousel/CarouselProgressBar.style.ts
