+++ src/components/home/Carousel/CarouselProgressBar.tsx
@@ -1,4 +1,4 @@
-import { Text } from "../__common__/typography/Text";
+import { Text } from "../../__common__/typography/Text";
 import {
     CarouselProgressBadge,
     CarouselProgressBarContainer,
